+++ src/main/scala/sc4pac/Downloader.scala
@@ -9,6 +9,7 @@ import coursier.cache as CC
 import zio.{ZIO, IO}
 import upickle.default as UP
 import sc4pac.JsonData as JD
+import sc4pac.error.Artifact2Error
 
 import Downloader.PartialDownloadSpec
 
@@ -16,6 +17,9 @@ import Downloader.PartialDownloadSpec
   * https://github.com/coursier/coursier/blob/8d93005b56dd84770c062aeae6d7a12c53948596/modules/cache/jvm/src/main/scala/coursier/cache/internal/Downloader.scala
   *
   * Our changes of the implementation resolve issues related to timeouts and resuming partial downloads.
+  *
+  * TODO This code is a mess and should be rewritten with less indirection and
+  * better wrapping of failures (e.g. using Either or ZIO).
   */
 class Downloader(
   artifact: Artifact,  // contains the URL
@@ -23,16 +27,16 @@ class Downloader(
   localFile: java.io.File,  // the local file after download
   logger: Logger,
   pool: java.util.concurrent.ExecutorService,
-  cookies: Downloader.Cookies,
+  credentials: Downloader.Credentials,
 ) {
 
-  def download: IO[CC.ArtifactError, java.io.File] = {
+  def download: IO[Artifact2Error, java.io.File] = {
     val url = artifact.url
     logger.checkingArtifact(url, artifact)
     if (url.startsWith("file:/")) {
       ZIO.attemptBlocking {
         if (localFile.exists()) Right(localFile)
-        else Left(new CC.ArtifactError.NotFound(localFile.toString))
+        else Left(new Artifact2Error.NotFound(localFile.toString))
       }.catchSome {
         case scala.util.control.NonFatal(e) => ZIO.succeed(Left(wrapDownloadError(e, url)))
       }.orDie.absolve
@@ -41,15 +45,15 @@ class Downloader(
     }
   }
 
-  private def wrapDownloadError(e: Throwable, url: String): CC.ArtifactError = e match {
-    case e: CC.ArtifactError.DownloadError => e
-    case _ => CC.ArtifactError.DownloadError(
+  private def wrapDownloadError(e: Throwable, url: String): Artifact2Error = e match {
+    case e: Artifact2Error => e
+    case _ => Artifact2Error.DownloadError(
       s"Caught ${e.getClass().getName()}${Option(e.getMessage).fold("")(" (" + _ + ")")} while downloading $url",
       Some(e)
     )
   }
 
-  private def remote(file: java.io.File, url: String): IO[CC.ArtifactError, Unit] = {
+  private def remote(file: java.io.File, url: String): IO[Artifact2Error, Unit] = {
     Downloader.attemptCancelableOnPool(pool, (isCanceled) => {
       val tmp = coursier.paths.CachePath.temporaryFile(file)  // file => .file.part
       logger.downloadingArtifact(url, artifact)
@@ -58,9 +62,9 @@ class Downloader(
         val res = downloading(url, file)(
           CC.CacheLocks.withLockOr(cacheLocation, file)(
             doDownload(file, url, tmp, isCanceled),
-            ifLocked = Some(Left(new CC.ArtifactError.Locked(file)))  // should not be an issue as long as just one instance of sc4pac is running
+            ifLocked = Some(Left(new Artifact2Error.Locked(file)))  // should not be an issue as long as just one instance of sc4pac is running
           ),
-          ifLocked = Some(Left(new CC.ArtifactError.Locked(file)))  // should not be an issue as long as just one instance of sc4pac is running
+          ifLocked = Some(Left(new Artifact2Error.Locked(file)))  // should not be an issue as long as just one instance of sc4pac is running
         )
         success = res.isRight
         res
@@ -74,20 +78,20 @@ class Downloader(
     * resumption attempts.
     */
   private def downloading[T](url: String, file: java.io.File)(
-    f: => Either[CC.ArtifactError, T], ifLocked: => Option[Either[CC.ArtifactError, T]]
-  ): Either[CC.ArtifactError, T] = {
+    f: => Either[Artifact2Error, T], ifLocked: => Option[Either[Artifact2Error, T]]
+  ): Either[Artifact2Error, T] = {
 
     @tailrec
-    def helper(retrySsl: Int, retryResumption: Int): Either[CC.ArtifactError, T] = {
+    def helper(retrySsl: Int, retryResumption: Int): Either[Artifact2Error, T] = {
       require(retrySsl >= 0 && retryResumption >= 0)
 
-      val resOpt: Either[(Int, Int), Either[CC.ArtifactError, T]] =
+      val resOpt: Either[(Int, Int), Either[Artifact2Error, T]] =
         try {
           val res0 = CC.CacheLocks.withUrlLock(url) {
             try f
             catch {
               case nfe: java.io.FileNotFoundException if nfe.getMessage != null =>
-                Left(new CC.ArtifactError.NotFound(nfe.getMessage))
+                Left(new Artifact2Error.NotFound(nfe.getMessage))
             }
           }
           res0.orElse(ifLocked).toRight((retrySsl - 1, retryResumption))  // as a safe-guard, we also decrease retry counter here
@@ -100,7 +104,7 @@ class Downloader(
         }
 
       resOpt match {
-        case Right(Left(ex: CC.ArtifactError.WrongLength)) if ex.got < ex.expected && retryResumption >= 1 =>
+        case Right(Left(ex: Artifact2Error.WrongLength)) if ex.got < ex.expected && retryResumption >= 1 =>
           System.err.println(s"File transmission incomplete (${ex.got}/${ex.expected}): trying to resume download $url")
           helper(retrySsl, retryResumption - 1)
         case Right(res) => res
@@ -111,21 +115,23 @@ class Downloader(
   }
 
   /** Download in blocking fashion. */
-  private def doDownload(file: java.io.File, url: String, tmp: java.io.File, isCanceled: AtomicBoolean): Either[CC.ArtifactError, Unit] = {
+  private def doDownload(file: java.io.File, url: String, tmp: java.io.File, isCanceled: AtomicBoolean): Either[Artifact2Error, Unit] = {
     var conn: URLConnection = null
 
     try {
-      val (conn0, partialDownload) = Downloader.urlConnectionMaybePartial(url, Downloader.PartialDownloadSpec.initBlocking(tmp), cookies)
+      val (conn0, partialDownload) = Downloader.urlConnectionMaybePartial(url, Downloader.PartialDownloadSpec.initBlocking(tmp), credentials)
       conn = conn0
 
       val respCodeOpt = CC.CacheUrl.responseCode(conn)
 
       if (respCodeOpt.contains(404))
-        Left(new CC.ArtifactError.NotFound(url, permanent = Some(true)))
+        Left(new Artifact2Error.NotFound(url))
       else if (respCodeOpt.contains(403))
-        Left(new CC.ArtifactError.Forbidden(url))
+        Left(new Artifact2Error.Forbidden(url))
       else if (respCodeOpt.contains(401))
-        Left(new CC.ArtifactError.Unauthorized(url, realm = CC.CacheUrl.realm(conn)))
+        Left(new Artifact2Error.Unauthorized(url))
+      else if (respCodeOpt.contains(429))
+        Left(new Artifact2Error.RateLimited(url))
       else {
         val lenOpt: Option[Long] =
           for (len0 <- Option(conn.getContentLengthLong).filter(_ >= 0L).orElse(Downloader.lengthFromContentRange(conn))) yield {
@@ -153,7 +159,7 @@ class Downloader(
               }
             }
 
-        def consumeStream(): Either[CC.ArtifactError, Unit] = {
+        def consumeStream(): Either[Artifact2Error, Unit] = {
           scala.util.Using.resource {
             val baseStream =
               if (conn.getContentEncoding == "gzip") new java.util.zip.GZIPInputStream(conn.getInputStream)
@@ -167,7 +173,7 @@ class Downloader(
             }
 
             if (!overlapRegionMatches) {
-              Left(new CC.ArtifactError.DownloadError(s"Partially downloaded file $tmp does not match remote file $url: delete the file and try again.", None))
+              Left(new Artifact2Error.DownloadError(s"Partially downloaded file $tmp does not match remote file $url: delete the file and try again.", None))
             } else {
               scala.util.Using.resource(
                 CC.CacheLocks.withStructureLock(cacheLocation) {
@@ -179,7 +185,7 @@ class Downloader(
                 if (interrupted) {
                   val msg = s"Download was canceled, so partially downloaded file is incomplete: $tmp"
                   logger.warn(msg)  // we log this directly, since the error cannot usually be handled via API after websocket was closed
-                  Left(new CC.ArtifactError.DownloadError(msg, None))
+                  Left(new Artifact2Error.DownloadError(msg, None))
                 } else {
                   Right(())
                 }
@@ -188,7 +194,7 @@ class Downloader(
           }
         }
 
-        def lengthCheck(): Either[CC.ArtifactError, Unit] =
+        def lengthCheck(): Either[Artifact2Error, Unit] =
           lenOpt match {
             case None => Right(())
             case Some(len) =>
@@ -196,7 +202,7 @@ class Downloader(
               if (len == tmpLen)
                 Right(())
               else
-                Left(new CC.ArtifactError.WrongLength(tmpLen, len, tmp.getAbsolutePath))
+                Left(new Artifact2Error.WrongLength(tmpLen, len, tmp.getAbsolutePath))
           }
 
         for {
@@ -245,8 +251,8 @@ class Downloader(
 
 object Downloader {
 
-  class Cookies(val simtropolisCookie: Option[String])
-  val emptyCookiesLayer = zio.ZLayer.succeed(Cookies(simtropolisCookie = None))
+  class Credentials(val simtropolisCookie: Option[String], val simtropolisToken: Option[String])
+  val emptyCredentialsLayer = zio.ZLayer.succeed(Credentials(simtropolisCookie = None, simtropolisToken = None))
 
   /** Returns true on success, false if data transfer was canceled. */
   private def readFullyTo(
@@ -358,7 +364,7 @@ object Downloader {
   /** Open a URL connection for download, optionally for resuming a partial
     * download (if byte-serving is supported by the server).
     */
-  private def urlConnectionMaybePartial(url0: String, specOpt: Option[PartialDownloadSpec], cookies: Downloader.Cookies): (URLConnection, Option[PartialDownloadSpec]) = {
+  private def urlConnectionMaybePartial(url0: String, specOpt: Option[PartialDownloadSpec], credentials: Downloader.Credentials): (URLConnection, Option[PartialDownloadSpec]) = {
 
     var conn: URLConnection = null
 
@@ -374,11 +380,15 @@ object Downloader {
             conn0.setConnectTimeout(Constants.urlConnectTimeout.toMillis.toInt)  // timeout for establishing a connection
             conn0.setReadTimeout(Constants.urlReadTimeout.toMillis.toInt)  // timeout in case of internet outage while downloading a file
 
-            // Set session cookie for rudimentary authentication to Simtropolis.
-            for (cookie <- cookies.simtropolisCookie) {
-              val host = conn0.getURL().getHost()
-              if (host == "simtropolis.com" || host.endsWith(".simtropolis.com")) {
-                conn0.setRequestProperty("Cookie", cookie)
+            val host = conn0.getURL().getHost()
+            if (host == "simtropolis.com" || host.endsWith(".simtropolis.com")) {
+              if (credentials.simtropolisToken.isDefined) {
+                conn0.addRequestProperty("Authorization", s"""SC4PAC-TOKEN-ST userkey="${credentials.simtropolisToken.get}"""")
+              } else {
+                // Set session cookie for rudimentary authentication to Simtropolis.
+                for (cookie <- credentials.simtropolisCookie) {
+                  conn0.setRequestProperty("Cookie", cookie)
+                }
               }
             }
           case _ =>
@@ -387,7 +397,17 @@ object Downloader {
         def makeResult[A](x: A): Right[A, (URLConnection, A)] = {
           if (is4xx(conn)) {
             closeConn(conn)
-            throw new Exception(s"Connection error 4xx: $conn")
+            val respCodeOpt = CC.CacheUrl.responseCode(conn)
+            if (respCodeOpt.contains(404))
+              throw new Artifact2Error.NotFound(url0)
+            else if (respCodeOpt.contains(403))
+              throw new Artifact2Error.Forbidden(url0)
+            else if (respCodeOpt.contains(401))
+              throw new Artifact2Error.Unauthorized(url0)
+            else if (respCodeOpt.contains(429))
+              throw new Artifact2Error.RateLimited(url0)
+            else
+              throw new Exception(s"Connection error ${respCodeOpt.map(_.toString).getOrElse("4xx")}: $conn")  // TODO use an Artifact2Error
           } else {
             Right((conn, x))
           }
@@ -415,7 +435,7 @@ object Downloader {
 
     res match {
       case Left(specOpt) =>
-        urlConnectionMaybePartial(url0, specOpt, cookies)  // reconnect, possibly starting from 0
+        urlConnectionMaybePartial(url0, specOpt, credentials)  // reconnect, possibly starting from 0
       case Right(ret) =>
         ret
     }